data_generation/requirements.txt
# Data Generation Dependencies
faker>=20.0.0
fhir.resources>=7.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
python-dateutil>=2.8.0
